Move cpu files
