+++ taichi/runtime/program_impls/llvm/llvm_program.cpp
@@ -6,7 +6,7 @@
 #include "taichi/codegen/llvm/struct_llvm.h"
 #include "taichi/runtime/llvm/llvm_offline_cache.h"
 #include "taichi/codegen/codegen.h"
-#include "taichi/backends/cpu/aot_module_builder_impl.h"
+#include "taichi/runtime/cpu/aot_module_builder_impl.h"
 
 #if defined(TI_WITH_CUDA)
 #include "taichi/runtime/cuda/aot_module_builder_impl.h"
